+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changed
 - Fixed AlphaBeta bug related to temporary atoms being removed too soon from atom2tokens.
+- Hypergraph.add_sequence() converts sequence name directly to atom.
 
 ### Removed